% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASCSMOTE.R
\name{ASC_SMOTE}
\alias{ASC_SMOTE}
\title{SMOTE - Imbalanced Data Treatment}
\usage{
ASC_SMOTE(form, data, small_prop = 0.5, seed = 8, k = 5, ...)
}
\arguments{
\item{form}{A formula describing the prediction problem}

\item{data}{A data frame containing the original (imbalanced) data set}

\item{small_prop}{A new desired proportion for minority class. Default 0.5}

\item{seed}{A single value, interpreted as an integer. Default 8}

\item{k}{A number indicating the number of nearest neighbours that are used to generate the new examples of the minority class. Default 5}
}
\description{
This function handles Imbalanced Data by using SMOTE method.
}
\examples{
\donttest{
## A small example with a data set created artificially from the IRIS
data(iris)
data <- iris[, c(1, 2, 5)]
data$Species <- factor(ifelse(data$Species == "setosa","rare","common"))

## checking the class distribution of this artificial data set
table(data$Species)

## now using SMOTE to create a more "balanced problem"
newData <- ASC_SMOTE(Species ~ ., data, small_prop = 0.5)
table(newData$Species)
}

}
